# Improve two_shot_all_reduce