Summary:
X-link: https://github.com/facebookresearch/FBGEMM/pull/231

- Adjust two_shot_all_reduce for better performance on AMD/NVIDIA GPUs
- Improve all_reduce time for 70B decoding, BS96
   - AMD MI300X: 35.14us to 32.87us
   - NVIDIA H100: 22.32us to 21us

Differential Revision: D62753553